fix: convert Amount and Balance to integer pence throughout pipeline

+++ schema_validators.py
@@ -3,9 +3,9 @@ import pandera.pandas as pa
 input_schema_validator = pa.DataFrameSchema(
     {
         "Date": pa.Column(pa.DateTime),
-        "Amount": pa.Column(float),
+        "Amount": pa.Column(int),
         "Desc": pa.Column(str, pa.Check.str_length(min_value=1)),
-        "Balance": pa.Column(float)
+        "Balance": pa.Column(int)
     },
     strict=True,
     coerce=False,
